Update TwitterSearchAlgorithm2.java
Sort array3 in descending order after we have created the HashMap from its values and the variable from our for loop. The loop that creates our HashMap only needs to be as long as our num variable because thats all we are outputting. Change long b = 0; to int b = 0; for faster output.

+++ TwitterSearchAlgorithm2.java
@@ -67,7 +67,7 @@ public class Search{
 		long a = 1;
 		long c = 0;
 		long d = 0;
-		for(long b = 0; b < num; b++)
+		for(int b = 0; b < num; b++)
 		{
 			//here we just call a String.valueOf() to skip assigning int array3 to an String array3
 			//multiple positions with the same key get assigned this way?
